v0.3.1143 follow-on — SCALE-SEAM ⓷ ⓸ ⓹, license / land / pins
Eleven methods out of client.ts (1,663 → 1,584). No extra withX().
